AbstractInterpreter: account for overlay possibility in unanalyzed call
When we skip inference of a call on `throw` block, previously we only
checked if the call is overlayed or not. But the call may call an
overlayed method internally, thus we need to conservatively taint the
`:nonoverlayed` bit when `interp` uses overlay method table.

Nevertheless this will not introduce any regressions on GPUCompiler
stack (like #48097), since it defines `InferenceParams(::GPUInterpreter)`
overload to turn off `unoptimize_throw_blocks`
<https://github.com/JuliaGPU/GPUCompiler.jl/blob/d5086fb3d93bbc4795a96f6f1457898af46a24cb/src/interface.jl#L272>