Replace tput calls with a terminfo parser
Implement a terminfo parser, and use it to replace the tput calls.

While tput calls are fairly safe, given tput it part of the UNIX
specification, it's nice to remove a system dependency and call one less
executable. More crucially, switching to parsing terminfo ourselves
makes it much more convenient to check terminfo capabilities. A
try-catch wrapped success(`tput termcap`) call can now be replaced with
a simple haskey(current_terminfo, :termcap) or get(current_terminfo,
:termcap, false) as appropriate.

This is part of the groundwork for styled printing of StyledStrings.