Buffer styled printing
When printing directly to stdout, there is a non-negligible overhead
compared to simply printing to an IOBuffer. Testing indicates 3
allocations per print argument, and benchmarks reveal a ~2x increase in
allocations overall and much as a 10x increase in execution time.

Thus, it seems worthwhile to use a temporary buffer in all cases.