Restore `modname == NULL` behavior for `jl_load_dynamic_library`
Within the Julia runtime, this usage has been superseded by the new
`jl_find_library_by_addr`, but this is still used downstream.

In particular, CBinding.jl ends up using this branch to get a handle
to libjulia-internal (perhaps a bit accidentally, since the comments
in CBinding.jl suggest it intends to get a handle to the exe)

This restores the behavior to avoid downstream breakage.