Add `WITH_TRACY_CALLSTACKS` build option
This change enables rudimentary call-stack support in Tracy. By default,
call stack sampling is turned on for all zones along with random
periodic sampling by the system.

In the future, we'll probably want to take a more fine-grained approach
to judiciously collect stack traces only where there are important, but
for now gives users a flag to play with to uncover more information.