Inline StepRange construction (#49270)
This can improve performance when iterating over a range with a step as currently steprange_last is not inlined. About a 12ns improvement

I don't think this should slow anything down as at worst you're making a call which is what you'd have to do without inlining. Also, it should be unlikely that branch is taken.
This should allow for better automatic inlining.