mention `Cmd(::Vector{String})` in `Cmd` docstring
This seems like a cyclical definition without it, and
naively `Cmd` and `String` seem on the surface to be
similar so it can be surprising that `Cmd("a string")`
does not work. Adding this signature in the docstring
should obviate the issues for first time users that read
the docstring without having reviewed the manual.

We also clarify the behavior of `exec` and command literals.