irinterp: Consider cfg information from discovered errors
If we infer a call to `Union{}`, we can terminate further abstract
interpretation. However, this of course also means that we can
make use of that information to refine the types of any phis
that may have originated from the basic block containing the
call that was refined to `Union{}`.