rework Profile stdlib and manual documentation
These docs seemed to have been developed disjointly,
and this commit reorganizes them for clarity and coherence.
This reworks them such that the tutorials and guides concerning
the `Profile` stdlib are housed in the `Profile` stdlib doc pages.
We leave the existing manual page to discuss the macros,
build options, and admonitions about the `Profile` std lib.