instrument llvm opt passes with timings
Right now we use a single name for all the LLVM passes with metadata being used to identify what the name of the pass run. In the future, we probably want to get rid of the hardcoded list of all possible zone names so that each pass can have its own zone.