macroexpand: stop pre-running the hygiene pass
This prepares us to delete the separate hygiene pass, and make it part
of lowering resolve-scopes. This affects a few macros which expect to
run macroexpand then mangle the result more, and the result of those
macros is not already wrapped in `esc()`. The damage seems relatively
minor however, and generally a good improvement.