more detailed docstrings for read! and write
The old `read!` docstring failed to describe how many bytes the
function reads and what its return value is.

The old `write` docstring didn't cover arrays.

I also documented `read!(..., Ref{T})`: I first wasn't sure if this
was meant to be documented, but then saw that it was already mentioned
in the docstring of `write`.