Don't depend on `CompilerSupportLibraries_jll` from `OpenBLAS_jll`
This is important because CSL_jll loads in many other libraries that we
may or may not care that much about, such as `libstdc++` and
`libgomp`.

We load `libstdc++` eagerly on Linux, so that will already be loaded in
all cases that we care about, however on macOS we don't generally want
that loaded, and this suppresses that.

`libgomp` is needed by BB-provided software that uses OpenMP during
compilation, however it can conflict with software compiled by the Intel
compilers, such as `MKL`.  It's best to allow MKL to load its OpenMP
libraries first, so delaying loading `libgomp` until someone actually
calls `using CompilerSupportLibraries_jll` is the right thing to do.

In the future, we want to rework JLLs such that libraries aren't eagerly
loaded at JLL `__init__()` time, but rather they should be JIT loaded
upon first usage of the library handle itself.  This would allow BB to
emit much more fine-grained dependency structures, so that the
distribution of a set of libraries can happen together, but the loading
of said libraries would be independent.