Improve `isassigned` implementation (#49827)
Unless `isassigned` is called on `Array` with `Int`s, it uses a try
catch, which is notoriously slow. This PR provides changes the default
implementation of `isassigned` to coerce the indices provided to `Int`s
and converts to linear or cartesian indices, depending on the arrays
`IndexStyle`. This also overloads `isassigned` for many of the array
types defined in Base.

Fixes: https://github.com/JuliaLang/julia/issues/44720