WIP: transition `@timeit` in Core.Compiler to use jl_timing instead
This builds on the prior work from Kristoffer and Prem to implement a
Julia-side `@zone` macro that interacts with the JL_TIMING system to
provide profiling events to Tracy, VTune, and/or the TIMING_COUNTS
back-end.

This implementation includes some extra tricks to make sure that the
timing block is stack-allocated, which can be important for performance
if we start to use `@zone` for high-frequency events.

Co-authored-by: Kristoffer Carlsson <kristoffer.carlsson@chalmers.se>
Co-authored-by: Prem Chintalapudi <prem.chintalapudi@gmail.com>