Prevent registering static FASTTLS with multiple libjulias
When there are multiple copies of libjulia present in the application,
they can both end up trying to use the same static fast tls storage.

The alternative options I considered for this were:
  - Add `jl_options.use_static_fasttls` and update PackageCompiler.jl to
    set the flag to FALSE
  - Use `dlvsym` in the CLI loader and attach to the exported
    jl_pgcstack_static symbols `__attribute__((symver(...)))`

The second approach is not available on musl, but the first one seems
like the ideal way to handle this long-term.