Disable prints. I think the extra allocs are coming from the guard check?
I'm not sure what it is, but the new code seems to be doing some kind of
type check on some value that it's loading. I'm not sure why, nor what
is different. I'll have to go back to the original code and change one
line at a time again to see.