FAQ: Clarify use of PROGRAM_FILE for importable scripts
In issue #45852, the proposed `@is_script` macro as a shorthand for
`abspath(PROGRAM_FILE) == @__FILE__` was rejected, because it was agreed that
the pattern of using the same file as a script and an importable library is
questionable, so there is no reason to make it too idiomatic.

For this reason, mention in the FAQ that while checking `PROGRAM_FILE` can be
used to check if a file is run as a script, it is not recommended to have files
double as both scripts and importable libraries.