Add check call to getrf! (#50134)
* Add check call to getrf!

`lu!(A; check=false)` is supposed to disable the checking and leave it to the user:

> When check = true, an error is thrown if the decomposition fails. When check = false, responsibility for checking the decomposition's validity (via issuccess) lies with the user.

However, this is not quite true since `lu!` calls `getrf!` which internally does a check for `chkfinite` which does throw an error. This updates the `getrf!` function to have a `check` argument which is then used by `lu!` to fully disable the error throwing checks.

* Update lapack.jl

(cherry picked from commit d69b1a228b4e0ac834c47a674725f0b08bd2da0e)