Fix refcount violations in new ifelse sroa lifting
Fixes a refcount violation in #49882. Unlike PhiNodes, the code
for ifelse was referencing the old arguments in the newly inserted
nodes, so they were getting counted. However, it did not update
the count when replacing these arguments later. Fix this by aligning
with PhiNode and leaving the arguments unset, pushing them in later.