Set default for TextTerminal
This PR adds defaults to the functions defined in `REPL.Terminals`. Specifically, it makes `Base.active_repl.t` the default terminal. The defaults are added in the generic fallbacks.

The branch name is related to the original thought behind this PR, but I could not get consistent results in testing for that. The idea was to use "\x1b[s" and "\x1b[u" to save and return to cursor positions. This worked initally, but then not, for reasons I could not understand. Help with implementing that would also be appreciated.