Add csave and crestore
After more testing, it figured out that the functions worked all along. I they just had no effect when I was saving and going to the final line in the terminal.

So this PR adds csave and crestore - the only functions with docstrings in this module. The use print instead of write, as returning the number of returned chars written is generally just noise, and provides no useful information, as it is the same return value each time the functions are run.