InteractiveUtils: add nvr (neovim-remote) to default editors
This patch adds [nvr (neovim-remote)][1] to the list of predefined
editors. With nvr as the editor files will open in an already running
neovim process, if one exist, without blocking the REPL. If no neovim
server is found, nvr will spawn neovim itself and block the REPL until
the process exits (just like when using nvim as the editor directly).

[1]: https://github.com/mhinz/neovim-remote