[DONOTMERGE] See if compilesig ever changes sparams
We have a check in here that checks whether changing to compilesig
changes the sparams and we refuse to use the compilesig in that
case (although the exact behavior changed recently). I'm not sure
the code path where we do change sparams is actually correct, but
I also couldn't come up with a test that actually goes down this
path. Putting in this assert shows that we don't hit it anywhere
in our existing testsuite either, so I'm putting this up as a PR
to pkgeval it and see if I'm missing anything. If not, we could
possibly remove typeintersect (or only do it with an assert in
debug mode).